Enable TLS on Cantor Server and the ability to create new users
- all users currently have full access
- unauthenticated users also have full access
- authenticated users have jwts stored in the backend for role management

+++ cantor-grpc-service/src/main/java/com/salesforce/cantor/grpc/ObjectsGrpcService.java
@@ -10,11 +10,10 @@ package com.salesforce.cantor.grpc;
 import com.google.protobuf.ByteString;
 import com.salesforce.cantor.Cantor;
 import com.salesforce.cantor.Objects;
+import com.salesforce.cantor.grpc.auth.*;
 import com.salesforce.cantor.grpc.objects.*;
 import io.grpc.Context;
 import io.grpc.stub.StreamObserver;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 import java.util.Collection;
@@ -53,6 +52,10 @@ public class ObjectsGrpcService extends ObjectsServiceGrpc.ObjectsServiceImplBas
             sendCancelledError(responseObserver, Context.current().cancellationCause());
             return;
         }
+        if (!writeRequestValid(request.getNamespace())) {
+            sendError(responseObserver, new UnauthorizedException("User not authorized to make this request: " + request));
+            return;
+        }
         try {
             getObjects().create(request.getNamespace());
             sendResponse(responseObserver, VoidResponse.getDefaultInstance());
@@ -67,6 +70,10 @@ public class ObjectsGrpcService extends ObjectsServiceGrpc.ObjectsServiceImplBas
             sendCancelledError(responseObserver, Context.current().cancellationCause());
             return;
         }
+        if (!writeRequestValid(request.getNamespace())) {
+            sendError(responseObserver, new UnauthorizedException("User not authorized to make this request: " + request));
+            return;
+        }
         try {
             getObjects().drop(request.getNamespace());
             sendResponse(responseObserver, VoidResponse.getDefaultInstance());
@@ -79,6 +86,10 @@ public class ObjectsGrpcService extends ObjectsServiceGrpc.ObjectsServiceImplBas
     public void keys(final KeysRequest request, final StreamObserver<KeysResponse> responseObserver) {
         if (Context.current().isCancelled()) {
             sendCancelledError(responseObserver, Context.current().cancellationCause());
+            return;
+        }
+        if (!readRequestValid(request.getNamespace())) {
+            sendError(responseObserver, new UnauthorizedException("User not authorized to make this request: " + request));
             return;
         }
         try {
@@ -100,6 +111,10 @@ public class ObjectsGrpcService extends ObjectsServiceGrpc.ObjectsServiceImplBas
             sendCancelledError(responseObserver, Context.current().cancellationCause());
             return;
         }
+        if (!readRequestValid(request.getNamespace())) {
+            sendError(responseObserver, new UnauthorizedException("User not authorized to make this request: " + request));
+            return;
+        }
         try {
             final GetResponse.Builder resultsBuilder = GetResponse.newBuilder();
             final byte[] value = getObjects()
@@ -119,6 +134,10 @@ public class ObjectsGrpcService extends ObjectsServiceGrpc.ObjectsServiceImplBas
             sendCancelledError(responseObserver, Context.current().cancellationCause());
             return;
         }
+        if (!writeRequestValid(request.getNamespace())) {
+            sendError(responseObserver, new UnauthorizedException("User not authorized to make this request: " + request));
+            return;
+        }
         try {
             getObjects().store(request.getNamespace(), request.getKey(), request.getValue().toByteArray());
             sendResponse(responseObserver, VoidResponse.getDefaultInstance());
@@ -133,6 +152,10 @@ public class ObjectsGrpcService extends ObjectsServiceGrpc.ObjectsServiceImplBas
             sendCancelledError(responseObserver, Context.current().cancellationCause());
             return;
         }
+        if (!writeRequestValid(request.getNamespace())) {
+            sendError(responseObserver, new UnauthorizedException("User not authorized to make this request: " + request));
+            return;
+        }
         try {
             final boolean result = getObjects().delete(request.getNamespace(), request.getKey());
             sendResponse(responseObserver, DeleteResponse.newBuilder().setResult(result).build());
@@ -145,6 +168,10 @@ public class ObjectsGrpcService extends ObjectsServiceGrpc.ObjectsServiceImplBas
     public void size(final SizeRequest request, final StreamObserver<SizeResponse> responseObserver) {
         if (Context.current().isCancelled()) {
             sendCancelledError(responseObserver, Context.current().cancellationCause());
+            return;
+        }
+        if (!readRequestValid(request.getNamespace())) {
+            sendError(responseObserver, new UnauthorizedException("User not authorized to make this request: " + request));
             return;
         }
         try {
